Split caldav requests in functions
